+++ src/verifier/verifier.rs
@@ -991,7 +991,6 @@ impl Verifier {
 
     /// Gets the package version meta, or `None` if the provider cannot be found.
     fn get_package_version_meta(&self, package_id: &PackageId, package: &InstalledPackageVersion) -> Result<Option<PackageVersionMeta>> {
-        // TODO: Refactor this, this shouldn't implemented here like this
         let repository = Repository::new(&package.metadata_repository_url, &package.metadata_repository_provider);
         let Some(provider) = create_metadata_provider(&repository) else {
             return Ok(None);